Split level into multiple files, really cleaning the code up.
